requirements.txt
coreapi>=2.3.1
Django>=1.11.3
django-filter>=1.0.4
djangorestframework>=3.6.3
gunicorn>=19.7.1
Markdown>=2.6.8
mysqlclient>=1.3.10
Pygments>=2.2.0
